Rename cap/floor pricers.
